person2/Code - Algorithm/requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
rasterio>=1.3.0
plotly>=5.18.0
matplotlib>=3.6.0
Pillow>=9.0.0
streamlit-image-coordinates>=0.1.6
